Add propTypes, adjust eslint, adjust App.js, adjust index.html title
