+++ .eslintrc
@@ -1,7 +1,12 @@
 {
     "parser": "babel-eslint",
     "plugins": ["react", "jest", "prettier", "markdown", "css-modules"],
-    "extends": ["airbnb", "plugin:prettier/recommended", "prettier/react", "plugin:css-modules/recommended"],
+    "extends": [
+        "airbnb",
+        "plugin:prettier/recommended",
+        "prettier/react",
+        "plugin:css-modules/recommended"
+    ],
     "globals": {
         "React": true,
         "document": true,
@@ -44,7 +49,7 @@
         "jsx-a11y/href-no-hash": "off",
         "jsx-a11y/anchor-is-valid": ["warn", { "aspects": ["invalidHref"] }],
         "react/jsx-boolean-value": ["warn", "never"],
-        "react/jsx-closing-bracket-location": ["warn", "after-props"],
+        "react/jsx-closing-bracket-location": [0, "tag-aligned"],
         "react/jsx-curly-spacing": ["warn", "never"],
         "react/jsx-filename-extension": ["warn", { "extensions": [".jsx"] }],
         "react/jsx-first-prop-new-line": ["warn", "multiline"],
