+++ package.json
@@ -59,7 +59,7 @@
         "bootstrap": "^4.4.1",
         "core-js": "3.6.4",
         "lodash": "4.17.15",
-        "prop-types": "15.7.2",
+        "prop-types": "^15.7.2",
         "react": "16.12.0",
         "react-bootstrap": "^1.0.0-beta.16",
         "react-dom": "16.12.0",
