+++ src/template.html
@@ -3,11 +3,10 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>Webpack react boilerplate</title>
+    <title>LEST Treecare</title>
 </head>
 <body>
 <noscript>
-    You need to enable JavaScript to run this app.
 </noscript>
 
 <div id="app"></div>
